cleared all errors but the result is fucked

+++ Scripts/Scenes/Play/2D/TouchFeedback/FeedbackSegment.cs
@@ -7,7 +7,7 @@ namespace WacK
     {
         [Export]
         private float LingerTime;
-        private float timer;
+        private double timer;
         private GEvents gEvents;
 
         public override void _Ready()
